refactor: move column and table_change into package rel
apache/gravitino#8589
